resolve circular conflict
